requirements.txt
fastapi
uvicorn
python-dotenv
pillow
pdf2image
google-generativeai
passlib[bcrypt]
pyjwt
psycopg2-binary
pydantic
sqlalchemy